+++ lib/src/directory_watcher/linux.dart
@@ -10,6 +10,8 @@ import 'dart:io';
 import 'package:async/async.dart';
 
 import '../directory_watcher.dart';
+import '../entity.dart';
+import '../fake_file_system_event.dart';
 import '../path_set.dart';
 import '../resubscribable.dart';
 import '../utils.dart';
@@ -19,7 +21,8 @@ import '../watch_event.dart';
 ///
 /// Inotify doesn't suport recursively watching subdirectories, nor does
 /// [Directory.watch] polyfill that functionality. This class polyfills it
-/// instead.
+/// instead. It also polyfills following symlinks, which inotify supports but
+/// isn't exposed via `Directory.watch`.
 ///
 /// This class also compensates for the non-inotify-specific issues of
 /// [Directory.watch] producing multiple events for a single logical action
@@ -53,10 +56,17 @@ class _LinuxDirectoryWatcher
   /// All known files recursively within [path].
   final PathSet _files;
 
-  /// [Directory.watch] streams for [path]'s subdirectories, indexed by name.
+  /// [Directory.watch] streams for [path]'s subdirectories, indexed by path.
   ///
   /// A stream is in this map if and only if it's also in [_nativeEvents].
   final _subdirStreams = <String, Stream<FileSystemEvent>>{};
+
+  /// [File.watch] streams for symlinks to files, indexed by path.
+  ///
+  /// These are needed because a directory doesn't emit an event if it contains
+  /// a symlink whose target file is modified. A stream is in this map if and
+  /// only if it's also in [_nativeEvents].
+  final _symlinkFileStreams = <String, Stream<FileSystemEvent>>{};
 
   /// A set of all subscriptions that this watcher subscribes to.
   ///
@@ -81,18 +91,12 @@ class _LinuxDirectoryWatcher
         onError: _eventsController.addError,
         onDone: _onDone);
 
-    _listen(new Directory(path).list(recursive: true), (entity) {
-      if (entity is Directory) {
-        _watchSubdir(entity.path);
-      } else {
-        _files.add(entity.path);
-      }
-    }, onError: (error, stackTrace) {
+    _checkContents(path, onError: (error, stackTrace) {
       _eventsController.addError(error, stackTrace);
       close();
     }, onDone: () {
       _readyCompleter.complete();
-    }, cancelOnError: true);
+    });
   }
 
   void close() {
@@ -102,13 +106,17 @@ class _LinuxDirectoryWatcher
 
     _subscriptions.clear();
     _subdirStreams.clear();
+    _symlinkFileStreams.clear();
     _files.clear();
     _nativeEvents.close();
     _eventsController.close();
   }
 
   /// Watch a subdirectory of [directory] for changes.
-  void _watchSubdir(String path) {
+  ///
+  /// If [isLink] is true, [path] is known to be a symlink and special behavior
+  /// is used to work around sdk#24815.
+  void _watchSubdir(String path, {bool isLink: false}) {
     // TODO(nweiz): Right now it's possible for the watcher to emit an event for
     // a file before the directory list is complete. This could lead to the user
     // seeing a MODIFY or REMOVE event for a file before they see an ADD event,
@@ -123,16 +131,41 @@ class _LinuxDirectoryWatcher
     // TODO(nweiz): Catch any errors here that indicate that the directory in
     // question doesn't exist and silently stop watching it instead of
     // propagating the errors.
-    var stream = new Directory(path).watch();
+
+    // TODO(nweiz): Gracefully handle the symlink edge cases described in the
+    // README. We could do so with some combination of watching containing
+    // directories and polling to see if nonexistent targets start to exist.
+
+    var stream;
+    if (isLink) {
+      // Work around sdk#24815 by listening to the concrete directory and
+      // post-processing the events so they have the paths we expect.
+      var resolvedPath = new Link(path).resolveSymbolicLinksSync();
+      stream = new Directory(resolvedPath).watch().map((event) =>
+          new FakeFileSystemEvent.rebase(event, resolvedPath, path));
+    } else {
+      stream = new Directory(path).watch();
+    }
+
     _subdirStreams[path] = stream;
+    _nativeEvents.add(stream);
+  }
+
+  /// Watch the target of a symlink at [path] that points to a file.
+  void _watchSymlink(String path) {
+    // Work around sdk#24815 by listening to the concrete file and
+    // post-processing the events so they have the paths we expect.
+    var resolvedPath = new Link(path).resolveSymbolicLinksSync();
+    var stream = new File(resolvedPath).watch().map((event) {
+      return new FakeFileSystemEvent.withPath(event, path);
+    });
+    _symlinkFileStreams[path] = stream;
     _nativeEvents.add(stream);
   }
 
   /// The callback that's run when a batch of changes comes in.
   void _onBatch(List<FileSystemEvent> batch) {
-    var files = new Set();
-    var dirs = new Set();
-    var changed = new Set();
+    var changes = <String, Entity>{};
 
     // inotify event batches are ordered by occurrence, so we treat them as a
     // log of what happened to a file. We only emit events based on the
@@ -144,86 +177,114 @@ class _LinuxDirectoryWatcher
       // stream is closed.
       if (event.path == path) continue;
 
-      changed.add(event.path);
-
+      var state = _stateFor(event);
       if (event is FileSystemMoveEvent) {
-        files.remove(event.path);
-        dirs.remove(event.path);
-
-        changed.add(event.destination);
-        if (event.isDirectory) {
-          files.remove(event.destination);
-          dirs.add(event.destination);
-        } else {
-          files.add(event.destination);
-          dirs.remove(event.destination);
-        }
-      } else if (event is FileSystemDeleteEvent) {
-        files.remove(event.path);
-        dirs.remove(event.path);
-      } else if (event.isDirectory) {
-        files.remove(event.path);
-        dirs.add(event.path);
+        changes[event.path] = new Entity.removed(path);
+        changes[event.destination] = state;
       } else {
-        files.add(event.path);
-        dirs.remove(event.path);
+        changes[event.path] = state;
       }
     }
 
-    _applyChanges(files, dirs, changed);
+    _applyChanges(changes);
   }
 
-  /// Applies the net changes computed for a batch.
+  /// Returns the [Entity] representing the result of [event].
   ///
-  /// The [files] and [dirs] sets contain the files and directories that now
-  /// exist, respectively. The [changed] set contains all files and directories
-  /// that have changed (including being removed), and so is a superset of
-  /// [files] and [dirs].
-  void _applyChanges(Set<String> files, Set<String> dirs, Set<String> changed) {
-    for (var path in changed) {
-      var stream = _subdirStreams.remove(path);
-      if (stream != null) _nativeEvents.add(stream);
+  /// This represents the current state of [event.path], except for a move
+  /// event, for which it represents the state of [event.destination].
+  Entity _stateFor(FileSystemEvent event) {
+    // Events that say they're directories are always non-symlink
+    // directories.
+    if (event.isDirectory) {
+      return new Entity(event.path, FileSystemEntityType.DIRECTORY);
+    }
+
+    // Delete events are simple, since whatever it was is gone now anyway.
+    if (event is FileSystemDeleteEvent) return new Entity.removed(event.path);
+
+    // For create, modify, and move events we need to check an actual entity
+    // on disk.
+    var path =
+        event is FileSystemMoveEvent ? event.destination : event.path;
+
+    // If it's not a link, then [event.isDirectory] is accurate. We
+    // checked it before, so here we know this must be a file.
+    if (!FileSystemEntity.isLinkSync(path)) {
+      return new Entity(path, FileSystemEntityType.FILE);
+    }
+
+    return new Entity(path, FileSystemEntity.typeSync(path), isLink: true);
+  }
+
+  /// Applies the net [changes] computed for a batch.
+  void _applyChanges(Map<String, Entity> changes) {
+    changes.forEach((path, state) {
+      var stream = _subdirStreams.remove(path) ??
+          _symlinkFileStreams.remove(path);
+      if (stream != null) _nativeEvents.remove(stream);
 
       // Unless [path] was a file and still is, emit REMOVE events for it or its
       // contents,
-      if (files.contains(path) && _files.contains(path)) continue;
-      for (var file in _files.remove(path)) {
-        _emit(ChangeType.REMOVE, file);
+      if (!state.isFile || !_files.contains(path)) {
+        for (var file in _files.remove(path)) {
+          _emit(ChangeType.REMOVE, file);
+        }
       }
-    }
 
-    for (var file in files) {
-      if (_files.contains(file)) {
-        _emit(ChangeType.MODIFY, file);
+      // If [path] is a directory, watch it and emit events for its contents.
+      if (state.isDirectory) {
+        _watchSubdir(path, isLink: state.isLink);
+
+        _checkContents(path, onFile: (entity) {
+          _emit(ChangeType.ADD, entity.path);
+        }, onError: (error, stackTrace) {
+          // Ignore an exception caused by the dir not existing. It's fine if it
+          // was added and then quickly removed.
+          if (error is FileSystemException) return;
+
+          _eventsController.addError(error, stackTrace);
+          close();
+        });
+        return;
+      }
+
+      // If [path] was removed or is a broken symlink, do nothing.
+      if (!state.isFile) return;
+
+      // If [path] is a valid symlink to a file, watch it because otherwise we
+      // won't get events for its contents changing.
+      if (state.isLink) _watchSymlink(path);
+
+      // Emit an event for [path] itself being changed or added.
+      if (_files.contains(path)) {
+        _emit(ChangeType.MODIFY, path);
       } else {
-        _emit(ChangeType.ADD, file);
-        _files.add(file);
+        _emit(ChangeType.ADD, path);
+        _files.add(path);
       }
-    }
-
-    for (var dir in dirs) {
-      _watchSubdir(dir);
-      _addSubdir(dir);
-    }
+    });
   }
 
-  /// Emits [ChangeType.ADD] events for the recursive contents of [path].
-  void _addSubdir(String path) {
-    _listen(new Directory(path).list(recursive: true), (entity) {
-      if (entity is Directory) {
-        _watchSubdir(entity.path);
+  /// Recursively adds the contents of [path] to [_files].
+  ///
+  /// This also watches any subdirectories or symlinked files in [path] for
+  /// further events.
+  ///
+  /// If [onFile] is passed, this calls it for every file it traverses. If
+  /// [onError] and/or [onDone] are passed, they're forwarded to
+  /// [Stream.listen].
+  void _checkContents(String path, {void onFile(Entity entity),
+      void onError(error, StackTrace stackTrace), void onDone()}) {
+    _listen(listDirThroughLinks(path), (entity) {
+      if (entity.type == FileSystemEntityType.DIRECTORY) {
+        _watchSubdir(entity.path, isLink: entity.isLink);
       } else {
         _files.add(entity.path);
-        _emit(ChangeType.ADD, entity.path);
+        if (entity.isLink) _watchSymlink(entity.path);
+        if (onFile != null) onFile(entity);
       }
-    }, onError: (error, stackTrace) {
-      // Ignore an exception caused by the dir not existing. It's fine if it
-      // was added and then quickly removed.
-      if (error is FileSystemException) return;
-
-      _eventsController.addError(error, stackTrace);
-      close();
-    }, cancelOnError: true);
+    }, onError: onError, onDone: onDone, cancelOnError: true);
   }
 
   /// Handles the underlying event stream closing, indicating that the directory