+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.9.8
+
+* Improve support for symlinks. Where possible, symlinks are now treated as
+  normal files. For caveats, see the README.
+
 # 0.9.7
 
 * Fix a bug in `FileWatcher` where events could be added after watchers were